Blender 2.79 (Constant Quality) Update
I haven't had time to iron out all the bugs, but this version of the Render Script should work with Blender 2.79 RC +
(This script should be used on 2.79 and higher. Please use the original script on blender 2.78c and lower.)

For those of you that don't know, starting with Blender 2.79, Blender now allows you to encode with Constant Bitrate or Constant Quality. This script will display the correct render info. I literally did a quick update in about 1 hour, tested it and it seemed to work. I am putting it out early so those of you that are using 2.79 have a way to use it with the render script.  Expect to find a bug. If you find any bugs, let me know. :)

Remember to call the new Script name video_editors_render_script_2.79_ALPHA.py instead of the original script name video_editors_render_script_v1.py

Have fun! ;)