Blender 2.8 (Default settings causing slow render)
I have performed some test with Blender 2.8 and noticed that using Blender 2.8's default settings are causing Blender 2.8 (with UI) render times to be 2 times as long as blender 2.79b (with UI). In addition, I am noticing that Blender 2.8 is automatically rendering across all the cores of my 4-core CPU - by default. Oddly, 2.8 is useing more processing power, yet it is rendering slower. 

So I decide that I would save my blender 2.79b video project and open it in Blender 2.8 to test the render. The result was that Blender 2.8 was rending the 2.79b blend file faster than blender 2.79b. (A 12 second render, on 2.79b, was taking 8 seconds on 2.8. This is what I want to see. :)

What this means:

There is a default setting in blender 2.8 that is actually slowing down VSE render times. I have gone line by line through the blender 2.8 settings, comparing them to 2.79b, and can't seem to figure out what setting is causing the slowdown problem.

The good news is that it looks like Multicore is working on blender 2.8 by default. But, so far, the only way to get it to work is to Save your project in blender 2.79b, then open the project and render with blender 2.8.

Note: You can continue to use the Universal Blender script on Blender 2.70 - 2.79b. Do not use it on 2.8 (Yet). 

Once we figure out what setting is causeing the slowdown in blender 2.8, we can finally be done with the script on future versions of blender.