Added 2.8 Beta compatibility
This version of the script will now do multicore rendering with Blender 2.8X.
This is a universal script that will work with 2.7x and higher.

Blender 2.8 Beta users:
When rendering with Blender 2.8, this script will use Blender 2.79 color management settings. (You can turn it off on line 206 of the script.) This was required because the new default color management settings cause the render times to be 3X longer than 2.79. When you change blender 2.8 to use the same color management settings, render times return to normal.

It is important to also note the WEBM is still broken in blender 2.8 Beta as of 5/11/2019. Once the issues with WEBM has been resolved, I will alter the script to add support.